+++ services/productsService.js
@@ -2,6 +2,9 @@ const Joi = require('@hapi/joi');
 
 const { findIngredientByName } = require('../models/ingredientsModels');
 const { createProduct, findAllProducts, findProductById, updateProducts, deleteProduct, insertImage } = require('../models/productsModel');
+const validateId = require('../utils/validIdMongoDB');
+const { calculateCost } = require('./productsCostsService');
+const { verifyAdmin } = require('./usersServices');
 
 const productSchema = Joi.object({
   name: Joi.string().required().not().empty(),
@@ -31,6 +34,7 @@ const verifyIngredients = async(ingredients) => {
 };
 
 const getProductById = async(id) => {
+  validateId(id, 'Product');
   const product = await findProductById(id);
   if(!product) {
     const error = { status: 404, message: 'Product not found' };
@@ -40,7 +44,13 @@ const getProductById = async(id) => {
   return product;
 }
 
-const newProduct = async(name, price, quantity, ingredients) => {
+const newProduct = async(body, user) => {
+  const { name, price, quantity, ingredients} = body;
+
+  validateProducts(body);
+  await verifyIngredients(ingredients);
+  await verifyAdmin(user);
+
   const productId = await createProduct(name, price, quantity, ingredients);
   const product = {
     _id: productId,
@@ -49,6 +59,8 @@ const newProduct = async(name, price, quantity, ingredients) => {
     quantity,
     ingredients
   };
+
+  await calculateCost(productId);
 
   return product;
 };
@@ -59,19 +71,35 @@ const getAllProducts = async() => {
   return result;
 };
 
-const editProduct = async(id, name, price, quantity, ingredients) => {
+const editProduct = async(id, user, body) => {
+  const { name, price, quantity, ingredients } = body;
+
+  validateId(id, 'Product');
+  await getProductById(id);
+  validateProducts(body);
+  await verifyIngredients(ingredients);
+  await verifyAdmin(user);
+
   await updateProducts(id, name, price, quantity, ingredients);
 
   return { message: `Product with id:${id} updated`}
 }
 
-const eraseProduct = async(id) => {
+const eraseProduct = async(id, user) => {
+  validateId(id, 'Product');
+  await getProductById(id);
+  await verifyAdmin(user)
+
   await deleteProduct(id);
 
   return { message: `Product with id:${id} deleted`}
 };
 
-const newImage = async(id, host) => {
+const newImage = async(id, host, user) => {
+  validateId(id, 'Product');
+  await getProductById(id);
+  await verifyAdmin(user)
+
   const image = `${host}/uploads/${id}.jpeg`;
   const { _id, name, price, quantity, ingredients } = await findProductById(id);
   await insertImage(id, image);
